Create README.md
This is a full stack library management system app developed with HTML, CSS and JS. We have used Express,js as the back-end framework application for the app that runs on top of Node.js. We have used MongoDB as the database application of our choice which is known for scalability, flexibility and popularity.